Move class "PerFrameResouce" from vulkan to class
