+++ class/GBufferPlanetMaterial.cpp
@@ -2,7 +2,6 @@
 #include "../vulkan/RenderPass.h"
 #include "../vulkan/GlobalDeviceObjects.h"
 #include "../vulkan/CommandBuffer.h"
-#include "../vulkan/PerFrameResource.h"
 #include "../vulkan/Framebuffer.h"
 #include "../vulkan/GlobalVulkanStates.h"
 #include "../vulkan/SharedIndirectBuffer.h"
@@ -17,6 +16,7 @@
 #include "RenderWorkManager.h"
 #include "FrameBufferDiction.h"
 #include "FrameWorkManager.h"
+#include "PerFrameResource.h"
 #include "../common/Util.h"
 #include "../vulkan/ShaderStorageBuffer.h"
 #include "../class/PlanetGeoDataManager.h"
