+++ class/Material.cpp
@@ -1,10 +1,8 @@
 #include <mutex>
 #include "Material.h"
 #include "../vulkan/CommandBuffer.h"
-#include "../vulkan/PerFrameResource.h"
 #include "../vulkan/PhysicalDevice.h"
 #include "../vulkan/GlobalDeviceObjects.h"
-#include "FrameWorkManager.h"
 #include "../vulkan/UniformBuffer.h"
 #include "../vulkan/SharedVertexBuffer.h"
 #include "../vulkan/SharedIndexBuffer.h"
@@ -18,21 +16,23 @@
 #include "../vulkan/ShaderModule.h"
 #include "../vulkan/Framebuffer.h"
 #include "../vulkan/DescriptorPool.h"
-#include "../class/MaterialInstance.h"
+#include "../vulkan/ComputePipeline.h"
 #include "../vulkan/ShaderStorageBuffer.h"
-#include "../class/UniformData.h"
 #include "../vulkan/CommandBuffer.h"
 #include "../vulkan/Image.h"
 #include "../vulkan/SharedIndirectBuffer.h"
-#include "RenderWorkManager.h"
 #include "../vulkan/GlobalVulkanStates.h"
 #include "../vulkan/GlobalDeviceObjects.h"
-#include "../class/PerMaterialIndirectUniforms.h"
+#include "UniformData.h"
+#include "MaterialInstance.h"
+#include "PerMaterialIndirectUniforms.h"
 #include "GlobalTextures.h"
-#include "../class/PerMaterialUniforms.h"
+#include "PerMaterialUniforms.h"
 #include "RenderPassBase.h"
-#include "../vulkan/ComputePipeline.h"
+#include "PerFrameResource.h"
 #include "Mesh.h"
+#include "RenderWorkManager.h"
+#include "FrameWorkManager.h"
 
 void Material::GeneralInit
 (
