+++ class/PostProcessingMaterial.cpp
@@ -2,7 +2,6 @@
 #include "../vulkan/RenderPass.h"
 #include "../vulkan/GlobalDeviceObjects.h"
 #include "../vulkan/CommandBuffer.h"
-#include "../vulkan/PerFrameResource.h"
 #include "../vulkan/Framebuffer.h"
 #include "../vulkan/GlobalVulkanStates.h"
 #include "../vulkan/SharedIndirectBuffer.h"
@@ -12,6 +11,7 @@
 #include "../vulkan/SwapChain.h"
 #include "../vulkan/DescriptorPool.h"
 #include "../vulkan/DescriptorSetLayout.h"
+#include "PerFrameResource.h"
 #include "RenderPassBase.h"
 #include "RenderPassDiction.h"
 #include "RenderWorkManager.h"
