+++ component/MeshRenderer.cpp
@@ -1,14 +1,10 @@
 #include "MeshRenderer.h"
-#include "../class/Mesh.h"
 #include "Material.h"
-#include "../class/MaterialInstance.h"
 #include <mutex>
 #include "../Base/BaseObject.h"
 #include "../vulkan/CommandBuffer.h"
-#include "../vulkan/PerFrameResource.h"
 #include "../vulkan/PhysicalDevice.h"
 #include "../vulkan/GlobalDeviceObjects.h"
-#include "../class/FrameWorkManager.h"
 #include "../vulkan/UniformBuffer.h"
 #include "../vulkan/SharedVertexBuffer.h"
 #include "../vulkan/SharedIndexBuffer.h"
@@ -16,13 +12,17 @@
 #include "../vulkan/GraphicPipeline.h"
 #include "../vulkan/SwapChain.h"
 #include "../vulkan/StagingBufferManager.h"
-#include "../class/RenderWorkManager.h"
 #include "../vulkan/GlobalVulkanStates.h"
-#include "../common/Singleton.h"
 #include "../vulkan/RenderPass.h"
 #include "../vulkan/Framebuffer.h"
 #include "../class/UniformData.h"
 #include "../class/Material.h"
+#include "../class/PerFrameResource.h"
+#include "../class/RenderWorkManager.h"
+#include "../class/FrameWorkManager.h"
+#include "../class/MaterialInstance.h"
+#include "../class/Mesh.h"
+#include "../common/Singleton.h"
 
 DEFINITE_CLASS_RTTI(MeshRenderer, BaseComponent);
