+++ vulkan/CommandPool.h
@@ -11,7 +11,7 @@ class CommandPool : public DeviceObjectBase<CommandPool>
 public:
 	~CommandPool();
 
-	bool Init(const std::shared_ptr<Device>& pDevice, uint32_t queueFamilyIndex, const std::shared_ptr<PerFrameResource>& pPerFrameRes, const std::shared_ptr<CommandPool>& pSelf, VkCommandPoolCreateFlags flags = 0);
+	bool Init(const std::shared_ptr<Device>& pDevice, uint32_t queueFamilyIndex, const std::shared_ptr<CommandPool>& pSelf, VkCommandPoolCreateFlags flags = 0);
 
 public:
 	std::shared_ptr<CommandBuffer> AllocatePrimaryCommandBuffer();
@@ -20,7 +20,6 @@ public:
 	std::vector<std::shared_ptr<CommandBuffer>> AllocateSecondaryCommandBuffers(uint32_t count);
 	VkCommandPool GetDeviceHandle() const { return m_commandPool; }
 	void Reset(const std::shared_ptr<Fence>& pFence);
-	std::weak_ptr<PerFrameResource> GetPerFrameResource() const { return m_pPerFrameRes; }
 	const VkCommandPoolCreateInfo& GetInfo() const { return m_info; }
 
 public:
@@ -31,5 +30,4 @@ public:
 protected:
 	VkCommandPool					m_commandPool;
 	VkCommandPoolCreateInfo			m_info;
-	std::weak_ptr<PerFrameResource> m_pPerFrameRes;
 };
