+++ vulkan/StagingBufferManager.cpp
@@ -4,7 +4,6 @@
 #include <algorithm>
 #include "Queue.h"
 #include "CommandBuffer.h"
-#include "PerFrameResource.h"
 
 bool StagingBufferManager::Init(const std::shared_ptr<Device>& pDevice, const std::shared_ptr<StagingBufferManager>& pSelf)
 {
